requirements.txt

Flask==3.1.1
smllib==1.7
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
azure-monitor-opentelemetry
Flask-Bootstrap4==4.0.2
